refactor the engine (it's too big)
